LayoutTests/networkinformation/window-property-expected.txt

Tests that the window.navigator.connection properties are present.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof navigator.webkitConnection == 'object' is true
PASS hasOnConnectionProperty() is true
PASS navigator.webkitConnection.hasOwnProperty('onwebkitnetworkinfochange') is true
PASS successfullyParsed is true

TEST COMPLETE
